Fixed the location of the first Qovax on the second part of Platforming Challenge Room
BUG: Qovax (and potentially all enemies, did not test) recognize the player with their raycast through platforms. This could cause a lot of trouble.
example: I had to move a Qovax that spawned near a respawn-platform because it immidatley targeted the player and tried to charge it through the platform.
While it couldn't go through it, the Qovax was now in the wrong locatio, which could hinder the player's ability to complete a platforming puzzle.